IA_agent/italienAgent/requirements.txt
mistralai
python-dotenv
gtts
pydub
weasyprint